Add controller tests for legacy contact endpoints
Currently, the legacy contact tests don't cover the controller. This
commit updates the service tests to run through the controller, and adds
a few extra tests to cover invalid request errors.

PER-9515
